+++ src/ast/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod captures;
 pub mod nodes;
+pub mod reachability;
 pub mod types;
 
 pub use nodes::*;